PlayerFighter: fix short-alignment pad at field_0x12d (-0x4 residual) — +3 linked_exact
After the redirect, a residual -4 (4 votes) remained: `short field_0x12d` forced 2-byte alignment,
inserting a pad at 0x12d that shifted field_0x130/route/boundingVolumes/trail/.. +4. Modeled the short
as two bytes (field_0x12d/field_0x12e) so 0x12c-0x130 packs to 4 bytes. parity 35/0/0, byte_exact 1001->
1002, linked_exact 1797->1800 (+3).

+++ src/game/ship/PlayerFighter.h
@@ -60,7 +60,9 @@ public:
     Explosion *explosion;
     int32_t field_0x128;
     uint8_t field_0x12c;
-    short field_0x12d;
+    // field_0x12d is a short@0x12d in the original (0x12d-0x12f); modeled as two bytes so the packed
+    // layout matches (a `short` here forces 2-byte alignment -> a pad at 0x12d that shifts field_0x130+).
+    uint8_t field_0x12d;
     uint8_t field_0x12e;
     signed char field_0x12f;
     int32_t field_0x130;
